+++ map/test.cub
@@ -4,7 +4,7 @@ SO ./texture/teapot.xpm
 WE ./texture/tails.xpm
 EA ./texture/planks.xpm
 
-F 46,139,87
+F 256,139,87
 C 135,206,235
 
  111111111
