+++ srcs/init_utils/parse_cubfile_utils/parse_fc_utils/get_rgb_color.c
@@ -6,107 +6,13 @@
 /*   By: kamitsui <kamitsui@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/16 18:17:19 by kamitsui          #+#    #+#             */
-/*   Updated: 2024/11/18 05:39:24 by kamitsui         ###   ########.fr       */
+/*   Updated: 2024/11/18 16:48:04 by kamitsui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-/**
- * @brief 
- *
- * @param str
- * @param error
- *
- * @return OK(0~255), Error: -1
- * @note error case:(str==NULL, empty string, over range, non-digit characters)
- *  Put error message of each case
- */
-static int	atoi_0_to_255(char *str, const char *entry, const char *rgb_str)
-{
-	int		result;
-	char	*str_num;
-	int		status;
-
-	result = 0;
-	status = 0;
-	if (str == NULL)
-		return (-1);
-	while (*str != '\0' && *str == ' ')
-		str++;
-	if (*str == '\0')
-	{
-		ft_dprintf(STDERR_FILENO, "%s%c: ",
-			ERR_PROMPT, *entry);
-		ft_dprintf(STDERR_FILENO, "%s\n", EMSG_RGB_EMPTY);
-		return (-1);
-	}
-	if (*str == ',')
-	{
-		ft_dprintf(STDERR_FILENO, "%s%c: %s %s\n",
-			ERR_PROMPT, *entry, EMSG_RGB_MISS, rgb_str);
-		return (-1);
-	}
-	str_num = str;
-	while (*str != '\0' && ft_isdigit(*str) == true)
-	{
-		while (result == 0 && *str == '0')
-			str++;
-		if (status != -1)
-			result = result * 10 + (*str - '0');
-		if (result > 255)
-			status = -1;
-		//if (result <= 255)
-		//	status = 0;
-		////printf("[%c] result[%d]", *str, result);
-		//else
-		//	status = -1;
-		str++;
-	}
-	if (*str != '\0' && *str != ',')
-	{
-		ft_dprintf(STDERR_FILENO, "%s%c: ",
-			ERR_PROMPT, *entry);
-		ft_dprintf(STDERR_FILENO, "\"%c\" %s\n", *str, EMSG_RGB_NOT_NUM);
-		return (-1);
-	}
-	if (status == -1)
-	{
-		ft_dprintf(STDERR_FILENO, "%s%c: ",
-			ERR_PROMPT, *entry);
-		*str = '\0';
-		ft_dprintf(STDERR_FILENO, "\"%s\" %s\n", str_num, EMSG_RGB_RANGE_OUT);
-		return (-1);
-	}
-//	while (*str != '\0' && *str == ' ')
-//		str++;
-//	if (*str != '\0')
-//		return (-1);
-	return (result);
-}
-
-void	debug_get_rgb_color(int fd, t_enum_fc type, int rgb[3], const char *msg)
-{
-	const char	*s[3] = {"r", "g", "b"};
-	int			i;
-
-	if (IS_DEBUG != true)
-		return ;
-	if (type == ENUM_F)
-		ft_dprintf(fd, ">>> func debug_get_rgb_color <<< ... call by %s\n", msg);
-	ft_dprintf(fd, "\tcolor[%d]", type);
-	i = 0;
-	while (i < 3)
-	{
-		ft_dprintf(fd, "\t%s[%d]", s[i], rgb[i]);
-		i++;
-	}
-	ft_dprintf(fd, "\n");
-	if (type == ENUM_C)
-		ft_dprintf(fd, "\n");
-}
-
-bool	is_rgb_miss(int i_rgb, char char_key, const char *rgb_str)
+static bool	is_rgb_miss(int i_rgb, char char_key, const char *rgb_str)
 {
 	if (i_rgb < 3)
 	{
@@ -116,6 +22,7 @@ bool	is_rgb_miss(int i_rgb, char char_key, const char *rgb_str)
 	}
 	return (true);
 }
+
 /**
  * @brief 
  *
